Move version bump logic into publish task
Since bumping versions is only really necessary when publishing
the affected logic is moved into the publish task and the
dedicated bump-task removed
